# Conflicts:
#	data_manager.py

+++ data_manager.py
@@ -1,19 +1,24 @@
 import database_common
 
-
+#QUESTIONS
 @database_common.connection_handler
-def get_all(cursor):
+def get_all_questions(cursor):
     cursor.execute("""
-                    SELECT * FROM question ORDER BY submission_time DESC;
+                    SELECT question.*, person.username FROM question 
+                    JOIN person ON question.user_id = person.id 
+                    ORDER BY question.submission_time DESC;
                    """)
     questions = cursor.fetchall()
     return questions
 
 
 @database_common.connection_handler
-def get_latest(cursor):
+def get_latest_questions(cursor):
     cursor.execute("""
-                    SELECT * FROM question ORDER BY submission_time DESC LIMIT 5;
+                    SELECT question.*, person.username FROM question 
+                    JOIN person ON question.user_id = person.id 
+                    ORDER BY question.submission_time DESC
+                    LIMIT 5;
                    """)
     questions = cursor.fetchall()
     return questions
@@ -22,8 +27,8 @@ def get_latest(cursor):
 
 
 @database_common.connection_handler
-def add_new_question(cursor, title, message, image, username):
-    cursor.execute("INSERT INTO question (title, message, image, username) VALUES (%s, %s, %s, %s)", (title, message, image, username))
+def add_new_question(cursor, title, message, image, user_id):
+    cursor.execute("INSERT INTO question (title, message, image, user_id) VALUES (%s, %s, %s, %s)", (title, message, image, user_id))
     cursor.execute('SELECT LASTVAL()')  # Some psycopg2 magic to get the latest inserted id
     latest_added_question_id = cursor.fetchone()['lastval']
     return latest_added_question_id
@@ -31,14 +36,17 @@ def add_new_question(cursor, title, message, image, username):
 
 @database_common.connection_handler
 def get_question_by_id(cursor, id):
-    cursor.execute("""SELECT * FROM question WHERE id = %(id)s;""", {'id': id})
+    cursor.execute("""
+                    SELECT question.*, person.username FROM question 
+                    JOIN person ON question.user_id = person.id 
+                    WHERE question.id = %(id)s;""", {'id': id})
     question = cursor.fetchone()
     return question
 
 
 @database_common.connection_handler
-def remove_by_id(cursor, id):
-    return cursor.execute("""DELETE FROM question WHERE id = %(id)s;""", {'id': id})
+def remove_question_by_id(cursor, id, session_user_id):
+    return cursor.execute("""DELETE FROM question WHERE id = %(id)s AND user_id = %(session_user_id)s;""", {'id': id, 'session_user_id': session_user_id})
 
 
 @database_common.connection_handler
@@ -51,28 +59,39 @@ def update_question(cursor, id, title, message, image):
 
 
 @database_common.connection_handler
-def add_comment(cursor, question_id, message, username):
-    return cursor.execute("INSERT INTO comment (question_id, message, username) VALUES (%s, %s, %s)", (question_id, message, username))
+def add_comment_to_question(cursor, question_id, message, user_id):
+    return cursor.execute("INSERT INTO comment (question_id, message, user_id) VALUES (%s, %s, %s)", (question_id, message, user_id))
 
 
 @database_common.connection_handler
-def get_comments(cursor, question_id):
-    cursor.execute("""SELECT * FROM comment WHERE question_id = %(question_id)s;""", {'question_id': question_id})
-    comment = cursor.fetchall()
-    return comment
+def vote_for_question(cursor, id):
+    return cursor.execute("""UPDATE question 
+                                SET vote_number = vote_number + 1
+                                WHERE id = %(id)s;""", {'id': id})
+
+
+@database_common.connection_handler
+def increment_question_views(cursor, id):
+    return cursor.execute("""UPDATE question 
+                                SET view_number = view_number + 1
+                                WHERE id = %(id)s;""", {'id': id})
 
 
 #ANSWERS
 @database_common.connection_handler
 def get_answers_by_question_id(cursor, id):
-    cursor.execute("""SELECT * FROM answer WHERE question_id = %(id)s ORDER BY vote_number DESC;""", {'id': id})
+    cursor.execute("""
+                    SELECT answer.*, person.username FROM answer 
+                    JOIN person ON answer.user_id = person.id 
+                    WHERE question_id = %(id)s 
+                    ORDER BY vote_number DESC;""", {'id': id})
     answers = cursor.fetchall()
     return answers
 
 
 @database_common.connection_handler
-def add_answer(cursor, question_id, message, image, username):
-    return cursor.execute("INSERT INTO answer (question_id, message, image, username) VALUES (%s, %s, %s, %s)", (question_id, message, image, username))
+def add_answer(cursor, question_id, message, image, user_id):
+    return cursor.execute("INSERT INTO answer (question_id, message, image, user_id) VALUES (%s, %s, %s, %s)", (question_id, message, image, user_id))
 
 
 @database_common.connection_handler
@@ -110,8 +129,10 @@ def remove_answer(cursor, id):
 
 @database_common.connection_handler
 def search(cursor, search_term):
-    cursor.execute("""SELECT * FROM question 
-                      WHERE message ILIKE """ "'%" + search_term + "%'" """ OR title ILIKE """ "'%" + search_term + "%'")
+    cursor.execute("""
+                    SELECT question.*, person.username FROM question 
+                    JOIN person ON question.user_id = person.id 
+                    WHERE message ILIKE """ "'%" + search_term + "%'" """ OR title ILIKE """ "'%" + search_term + "%'")
     result = cursor.fetchall()
     return result
 
@@ -134,6 +155,7 @@ def get_question_id_by_comment_id(cursor, id):
     question_id = cursor.fetchone()['question_id']
     return question_id
 
+
 @database_common.connection_handler
 def get_answer_id_by_comment_id(cursor, id):
     cursor.execute("""SELECT answer_id FROM comment WHERE id = %(id)s;""", {'id': id})
@@ -142,33 +164,36 @@ def get_answer_id_by_comment_id(cursor, id):
 
 
 @database_common.connection_handler
-def increment_view(cursor, id):
-    return cursor.execute("""UPDATE question 
-                                SET view_number = view_number + 1
-                                WHERE id = %(id)s;""", {'id': id})
+def add_comment_answer(cursor, answer_id, message, user_id):
+    return cursor.execute("INSERT INTO comment (answer_id, message, user_id) VALUES (%s, %s, %s)", (answer_id, message, user_id))
 
 
 @database_common.connection_handler
-def add_comment_answer(cursor, answer_id, message, username):
-    return cursor.execute("INSERT INTO comment (answer_id, message, username) VALUES (%s, %s, %s)", (answer_id, message, username))
+def get_question_comments(cursor, question_id):
+    cursor.execute("""
+                    SELECT comment.*, person.username FROM comment 
+                    JOIN person ON comment.user_id = person.id 
+                    WHERE question_id = %(question_id)s;""", {'question_id': question_id})
+    comment = cursor.fetchall()
+    return comment
 
 
 @database_common.connection_handler
 def get_answer_comments(cursor, answer_id):
-    cursor.execute("""SELECT * FROM comment WHERE answer_id = %(answer_id)s;""", {'answer_id': answer_id})
+    cursor.execute("""
+                    SELECT comment.*, person.username FROM comment 
+                    JOIN person ON comment.user_id = person.id 
+                    WHERE answer_id = %(answer_id)s;""", {'answer_id': answer_id})
     comment = cursor.fetchall()
     return comment
 
+
 #USER
-
-
 @database_common.connection_handler
 def register_user(cursor, username, password):
     cursor.execute("INSERT INTO person (username, password) VALUES (%s, %s)", (username, password))
     person = cursor.fetchall()
     return person
-
-
 
 
 @database_common.connection_handler
@@ -185,11 +210,12 @@ def list_users(cursor):
     users = cursor.fetchall()
     return users
 
+
 @database_common.connection_handler
 def get_user(cursor, username):
     cursor.execute("SELECT * FROM person WHERE username = %(username)s ", {'username': username})
-    person = cursor.fetchone()
-    return person
+    user = cursor.fetchone()
+    return user
 
 
 @database_common.connection_handler
